fix: prediction_pick_points 인자를 bigint로 (rank() 반환형 일치)
중간 점검 SQL 정적 검토에서 발견: rank() 윈도우는 bigint를 반환하는데
함수 인자가 integer라 bigint→integer 암묵 캐스트가 없어 view 생성이
"function does not exist"로 실패했을 문제. 인자를 bigint로 바꿔 해결
(integer 값도 자동 확장돼 들어감).

+++ supabase/rollback/revert_toon_rank_scoring.sql
@@ -53,5 +53,5 @@ where f.finished
         = date_trunc('week', f.kickoff_at  at time zone 'Asia/Seoul')
   );
 
--- 새 integer 오버로드 제거
-drop function if exists public.prediction_pick_points(integer);
+-- 새 bigint 오버로드 제거
+drop function if exists public.prediction_pick_points(bigint);
